@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class TopBarActionButton extends StatelessWidget {
-  const TopBarActionButton({Key key}) : super(key: key);
+  const TopBarActionButton({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
